# docs/swagger.yaml
definitions:
...
  handlers.Cat:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /cats:
    post:
      consumes:
      - application/json
      description: AddCat
      parameters:
      - description: cat
        in: body
        name: cat
        required: true
        schema:
          $ref: '#/definitions/handlers.Cat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: AddCat
  /cats/{data}:
    get:
      consumes:
      - application/json
      description: GetCats
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      - description: type
        in: query
        name: type
        required: true
        type: string
      - description: data
        in: path
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetCats
...
swagger: "2.0"